Initial commit of IRC chat functionality. The basics now work.

+++ src/qt/coincontroldialog.cpp
@@ -712,7 +712,7 @@ void CoinControlDialog::updateView()
 
             // coin age
             int nDayWeight = (min((GetAdjustedTime() - out.tx->GetTxTime()), (int64) nStakeMaxAge) - nStakeMinAge) / 86400;
-            uint64 coinAge = max(out.tx->vout[out.i].nValue * nDayWeight / COIN, (int64) 0);
+            double coinAge = max(out.tx->vout[out.i].nValue * (double)nDayWeight / COIN, (double) 0);
             itemOutput->setText(COLUMN_COINAGE, strPad(QString::number(coinAge), 8, " "));
 
             // priority
